A different approach to the extension host problem
I think the work to figure out the extension host was much
overcomplicated. I don't think we need all that.
