+++ src/modules/cmdpal/Microsoft.CmdPal.Core.ViewModels/PageViewModelFactory.cs
@@ -15,7 +15,7 @@ public class PageViewModelFactory : IPageViewModelFactoryService
         _scheduler = scheduler;
     }
 
-    public PageViewModel? TryCreatePageViewModel(IPage page, bool nested, CommandPaletteHost host)
+    public PageViewModel? TryCreatePageViewModel(IPage page, bool nested, AppExtensionHost host)
     {
         return page switch
         {
